// src/typechain/Autonity.ts
export declare namespace Autonity {
  export type ValidatorStruct = {
    treasury: AddressLike;
    nodeAddress: AddressLike;
    oracleAddress: AddressLike;
    enode: string;
    commissionRate: BigNumberish;
    bondedStake: BigNumberish;
    unbondingStake: BigNumberish;
    unbondingShares: BigNumberish;
    selfBondedStake: BigNumberish;
    selfUnbondingStake: BigNumberish;
    selfUnbondingShares: BigNumberish;
    selfUnbondingStakeLocked: BigNumberish;
    liquidContract: AddressLike;
    liquidSupply: BigNumberish;
    registrationBlock: BigNumberish;
    totalSlashed: BigNumberish;
    jailReleaseBlock: BigNumberish;
    provableFaultCount: BigNumberish;
    consensusKey: BytesLike;
    state: BigNumberish;
  };

  export type ValidatorStructOutput = [
    treasury: string,
    nodeAddress: string,
    oracleAddress: string,
    enode: string,
    commissionRate: bigint,
    bondedStake: bigint,
    unbondingStake: bigint,
    unbondingShares: bigint,
    selfBondedStake: bigint,
    selfUnbondingStake: bigint,
    selfUnbondingShares: bigint,
    selfUnbondingStakeLocked: bigint,
    liquidContract: string,
    liquidSupply: bigint,
    registrationBlock: bigint,
    totalSlashed: bigint,
    jailReleaseBlock: bigint,
    provableFaultCount: bigint,
    consensusKey: string,
    state: bigint
  ] & {
    treasury: string;
    nodeAddress: string;
    oracleAddress: string;
    enode: string;
    commissionRate: bigint;
    bondedStake: bigint;
    unbondingStake: bigint;
    unbondingShares: bigint;
    selfBondedStake: bigint;
    selfUnbondingStake: bigint;
    selfUnbondingShares: bigint;
    selfUnbondingStakeLocked: bigint;
    liquidContract: string;
    liquidSupply: bigint;
    registrationBlock: bigint;
    totalSlashed: bigint;
    jailReleaseBlock: bigint;
    provableFaultCount: bigint;
    consensusKey: string;
    state: bigint;
  };

  export type PolicyStruct = {
    treasuryFee: BigNumberish;
    minBaseFee: BigNumberish;
    delegationRate: BigNumberish;
    unbondingPeriod: BigNumberish;
    initialInflationReserve: BigNumberish;
    treasuryAccount: AddressLike;
  };

  export type PolicyStructOutput = [
    treasuryFee: bigint,
    minBaseFee: bigint,
    delegationRate: bigint,
    unbondingPeriod: bigint,
    initialInflationReserve: bigint,
    treasuryAccount: string
  ] & {
    treasuryFee: bigint;
    minBaseFee: bigint;
    delegationRate: bigint;
    unbondingPeriod: bigint;
    initialInflationReserve: bigint;
    treasuryAccount: string;
  };

  export type ContractsStruct = {
    accountabilityContract: AddressLike;
    oracleContract: AddressLike;
    acuContract: AddressLike;
    supplyControlContract: AddressLike;
    stabilizationContract: AddressLike;
    upgradeManagerContract: AddressLike;
    inflationControllerContract: AddressLike;
    nonStakableVestingContract: AddressLike;
  };

  export type ContractsStructOutput = [
    accountabilityContract: string,
    oracleContract: string,
    acuContract: string,
    supplyControlContract: string,
    stabilizationContract: string,
    upgradeManagerContract: string,
    inflationControllerContract: string,
    nonStakableVestingContract: string
  ] & {
    accountabilityContract: string;
    oracleContract: string;
    acuContract: string;
    supplyControlContract: string;
    stabilizationContract: string;
    upgradeManagerContract: string;
    inflationControllerContract: string;
    nonStakableVestingContract: string;
  };

  export type ProtocolStruct = {
    operatorAccount: AddressLike;
    epochPeriod: BigNumberish;
    blockPeriod: BigNumberish;
    committeeSize: BigNumberish;
  };

  export type ProtocolStructOutput = [
    operatorAccount: string,
    epochPeriod: bigint,
    blockPeriod: bigint,
    committeeSize: bigint
  ] & {
    operatorAccount: string;
    epochPeriod: bigint;
    blockPeriod: bigint;
    committeeSize: bigint;
  };

  export type ConfigStruct = {
    policy: Autonity.PolicyStruct;
    contracts: Autonity.ContractsStruct;
    protocol: Autonity.ProtocolStruct;
    contractVersion: BigNumberish;
  };

  export type ConfigStructOutput = [
    policy: Autonity.PolicyStructOutput,
    contracts: Autonity.ContractsStructOutput,
    protocol: Autonity.ProtocolStructOutput,
    contractVersion: bigint
  ] & {
    policy: Autonity.PolicyStructOutput;
    contracts: Autonity.ContractsStructOutput;
    protocol: Autonity.ProtocolStructOutput;
    contractVersion: bigint;
  };

  export type CommitteeMemberStruct = {
    addr: AddressLike;
    votingPower: BigNumberish;
    consensusKey: BytesLike;
  };

  export type CommitteeMemberStructOutput = [
    addr: string,
    votingPower: bigint,
    consensusKey: string
  ] & { addr: string; votingPower: bigint; consensusKey: string };
}